Filter tab added.
1 .FilterViewController added.
2. Problems with constraints solved.
