+++ app/Conference/Conference/DayViewController.swift
@@ -11,7 +11,6 @@ import XLPagerTabStrip
 
 class DayViewController:UIViewController, IndicatorInfoProvider{
 	@IBOutlet weak var tableView: UITableView!
-	@IBOutlet weak var uiview: UIView!
 	
 	override func viewDidLoad() {
 		self.tableView.backgroundColor = UIColor(red: 0, green: 0, blue: 0, alpha: 0.05)
@@ -22,6 +21,7 @@ class DayViewController:UIViewController, IndicatorInfoProvider{
 		return IndicatorInfo(title: "Day 1")
 	}
 }
+
 extension DayViewController: UITableViewDelegate, UITableViewDataSource{
 
 	func numberOfSectionsInTableView(tableView: UITableView) -> Int {
@@ -29,7 +29,7 @@ extension DayViewController: UITableViewDelegate, UITableViewDataSource{
 	}
 	
 	func tableView(_ tableView: UITableView, numberOfRowsInSection section:Int) -> Int {
-		return 20
+		return 10
 	}
 	
 	func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
@@ -43,18 +43,4 @@ extension DayViewController: UITableViewDelegate, UITableViewDataSource{
 
 }
 
-extension UIView {
-	
-	func dropShadow(scale: Bool = true) {
-		
-		self.layer.masksToBounds = false
-		self.layer.shadowColor = UIColor.black.cgColor
-		self.layer.shadowOpacity = 0.18
-		self.layer.shadowOffset = CGSize(width: 0, height: 0)
-		self.layer.shadowRadius = 4
-		
-		self.layer.shadowPath = UIBezierPath(rect: self.bounds).cgPath
-		self.layer.shouldRasterize = true
-		self.layer.rasterizationScale = scale ? UIScreen.main.scale : 1
-	}
-}
+
