+++ app/Conference/Conference/FilterTableViewCell.swift
@@ -1,0 +1,15 @@
+//
+//  FilterTableViewCell.swift
+//  Conference
+//
+//  Created by SKIXY-MACBOOK on 07/06/17.
+//  Copyright © 2017 shubhamrathi. All rights reserved.
+//
+
+import UIKit
+
+class FilterTableViewCell: UITableViewCell{
+	
+	@IBOutlet weak var filterLabel: UILabel!
+	@IBOutlet weak var filterImage: UIImageView!
+}
