--- app/Conference/Conference/SecondViewController.swift
@@ -1,25 +1,0 @@
-//
-//  SecondViewController.swift
-//  Conference
-//
-//  Created by SKIXY-MACBOOK on 28/05/17.
-//  Copyright © 2017 shubhamrathi. All rights reserved.
-//
-
-import UIKit
-
-class SecondViewController: UIViewController {
-
-	override func viewDidLoad() {
-		super.viewDidLoad()
-		// Do any additional setup after loading the view, typically from a nib.
-	}
-
-	override func didReceiveMemoryWarning() {
-		super.didReceiveMemoryWarning()
-		// Dispose of any resources that can be recreated.
-	}
-
-
-}
-
